models/vlm/requirements.txt
kserve>=0.12.0
python-multipart>=0.0.6
# torch and torchvision installed separately with CUDA index in Dockerfile
transformers>=4.57.0
pillow>=10.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
qwen-vl-utils>=0.0.14
